Rename dg_camera.c
