+++ include/dg_camera.h
